requirements.txt
Flask
werkzeug
python-dotenv
pymongo
google-generativeai
gunicorn

